[backend] fused 8-byte rc box header: i32 count + i32 tag in one word
Shared/regional variant records now carry their box header themselves:
the layout is {4-byte count slot, i32 tag, payload} and a non-regional
rc box of such a record IS the record — the refcount overlays the
leading slot, so the rc pointer equals the record pointer and borrow is
the identity. [value] variants (never a box element; shared records
embed as pointers) keep the compact minimal-width-tag layout from the
packing change. Every other box shrinks its header too: the count is
i32 everywhere, so a compound box is {i32 count, elem at
alignTo(4, align)}.

Lowering changes: rc.inc/dec/fetch/set/is_unique/assume_unique operate
on i32 counts (index-typed SSA values widen/narrow at the boundary,
conditionally for 32-bit targets); rc.create stores the element FIRST
and the count second, since a whole-value store would clobber the
overlaid count slot; record.tag/coerce/variant and the fused creates
address tag/payload past the count slot on fused-header records; the
tagged-immediate dummy box becomes the literal fused header
{i32 count, i32 tag} (immortal threshold at width 32); the region
scanner advances onto the tag before the variant read. The runtime's
RcBox mirrors the i32 count, and the repl's reflection follows the new
layout rules.

Benchmarks (aarch64, n=4.2M, xLTO + static mimalloc; the harness
rbtree keys were aligned to Koka's int32 for an apples-to-apples
comparison):

- rbtree: 32-byte nodes. LLd misses 4.219M vs Koka's 4.225M — EXACT
  parity, two nodes per cache line as the line-footprint model
  predicted; RSS 133.4MB vs Koka 133.9MB — parity; wall 0.43s vs
  0.38s (1.11x, from 1.18x).
- nbe-hoas: Value boxes 40->24B, Term/Env 40->32B; RSS 386->252MB
  (-35%); LLd misses 14.9M -> 10.1M; wall 0.235 -> 0.21s (1.27x,
  from 1.52x).
- zipper: 0.40s.

Full lit suite (288, incl. asan/lsan/msan/tsan executing gates),
rrc/jit/repl/rt tests green; wasm32 driver test covers the 32-bit
index-width adapters.

+++ crates/reussir-rt/src/rc.rs
@@ -6,7 +6,9 @@ use std::{
 
 #[repr(C)]
 struct RcBox<T> {
-    count: Cell<usize>,
+    // Mirrors the compiler's box header: an i32 refcount, with the payload
+    // at `align_to(4, align(T))` (repr(C) gives exactly that).
+    count: Cell<u32>,
     data: UnsafeCell<T>,
 }
 
@@ -44,7 +46,7 @@ impl<T> Rc<T> {
     pub unsafe fn data_mut(&mut self) -> &mut T {
         unsafe { &mut *self.get_box().as_mut().data.get() }
     }
-    pub fn count_ref(&self) -> &Cell<usize> {
+    pub fn count_ref(&self) -> &Cell<u32> {
         unsafe { &self.get_box().as_ref().count }
     }
     pub fn is_unique(&self) -> bool {
